Can't get a commit msg from a manual tigger
I was trying to get the commit msg from a manual workflow trigger. Of course that does not have a commit msg. Now retrieving the last commit msg from the repo, this will either be the commit that triggered the test or the latest one pulled for a manual trigger. Also added a bunch of error handling